requirements.txt
Django>=5.2.2
Pillow>=10.0.0
django-ckeditor>=6.7.0
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0